Adicionando as regras e as validações

+++ modulo4/query-builders/src/endpoints/deletePerfumes.ts
@@ -1,11 +1,19 @@
 import { Request, Response } from "express";
 import connection from "../database/connection";
 import { TABLE_PERFUMES } from "../database/tableNames";
+import { addLeadingZero } from "../utils/addLeadingZero";
 
 export const deletePerfumes = async(req: Request, res: Response) => {
     let errorCode = 400
     try {
-        const id = req.params.id
+        const id = addLeadingZero(req.params.id)
+        const checkExists = await connection(TABLE_PERFUMES)
+            .select()
+            .where({ id : id})
+        if(checkExists.length === 0) {
+            errorCode = 404
+            throw new Error("Perfume não encontrado.")
+        }
         await connection(TABLE_PERFUMES)
             .delete()
             .where({ id: id })
